Add docs `DatetimeLocal::class`.
